Finish splitting up tests into modules
- Add remaining models tests to models module
- Create settings module
- Reformat conftest
