+++ tests/penn_chime/test_settings.py
@@ -1,0 +1,3 @@
+def test_defaults_repr(DEFAULTS):
+    repr(DEFAULTS)
+    # TODO: Add assertions here
